+++ legends/stats.py
@@ -3,9 +3,53 @@
 """
 
 from legends.utils.functions import formatDict
+from legends.utils.eventhandler import Event, EventHandler
 from legends.constants import PART_EFFECTS, POWER_GRADIENT, STAT_ABBREVIATIONS
 
-__all__ = ['PartEffects', 'StatObject', 'Stats']
+__all__ = [
+    'checkForStats',
+    'PartEffects',
+    'StatChangeEvent',
+    'StatObject',
+    'Stats'
+]
+
+def checkForStats(obj):
+    """Checks if the given objects has a 'stats' attribute that points
+    to a `StatObject` instance.
+
+    Args:
+        obj (obj): The object to check.
+
+    Raises:
+        ValueError: If the object fails the check.
+
+    """
+    if not hasattr(obj, 'stats'):
+        raise ValueError('{} has no `stats` attribute'.format(obj))
+    if not isinstance(obj.stats, StatObject):
+        raise ValueError(
+            '{} is not an instance of `StatObject`'.format(obj.stats)
+        )
+
+class StatChangeEvent(Event): # pylint: disable=too-few-public-methods
+    """Created when one of a `StatObject`'s attributes changes.
+
+    Should only be created when the changing attribute is a statistic.
+
+    Attributes:
+        parent (obj): The object whose statistics have changed.
+        attrName (str): The name of the statistic that changed.
+        oldVal (int or float): The old value of the statistic.
+        newVal (int or float): The new value of the statistic.
+
+    """
+
+    def __init__(self, parent, attrName, oldVal, newVal):
+        self.parent = parent
+        self.attrName = attrName
+        self.oldVal = oldVal
+        self.newVal = newVal
 
 class StatObject():
     """An object that stores named statistics as attributes.
@@ -22,58 +66,124 @@ class StatObject():
     the statistics of the given instances.
 
     Attributes:
-        statAbbrs (iterable of str): Strings that represent the stored
-            statistics. Typically abbreviations. Used to construct
-            object attributes.
+        parent (obj): The object to which these stats belong.
+        onChange (legends.utils.eventhandler.EventHandler): Sends a
+            `StatChangeEvent` to subscribers when a statistic changes.
+        silent (bool): Set to `True` to prevent the sending of a
+            `StatChangeEvent` when a statistic changes. Defaults to
+            `False`.
 
     """
 
-    def __init__(self):
-        self.statAbbrs = []
+    def __init__(self, statAbbrs, parent=None, initDict=None):
+        """If a dictionary of stat values is provided, the constructor
+        initializes the instance with these values.
+
+        Args:
+            statAbbrs (dict): {`str`:`str`} The dictionary to assign to
+                the `statAbbrs` property.
+            initDict (dict): {`str`:`int or float`} A dictionary mapping
+                stat names to numerical values. The keys of `initDict`
+                should match the keys of the `statAbbrs` property.
+
+        """
+        self._statAbbrs = statAbbrs
+        self.parent = parent
+        self.silent = False
+        self.onChange = EventHandler()
+        if initDict is None:
+            initDict = {statName: 0 for statName in self.statAbbrs}
+        self.update(initDict)
+
+    @property
+    def statAbbrs(self):
+        """dict: {`str`:`str`} A dictionary mapping the names of stats
+        to be stored to the abbreviations that are to be used as
+        attribute names.
+        """
+        return self._statAbbrs
+
+    def __setattr__(self, attrName, value):
+        oldVal = getattr(self, attrName) if hasattr(self, attrName) else None
+        self.__dict__[attrName] = value
+        if (
+            attrName in self.statAbbrs.values()
+            and oldVal != value
+            and not self.silent
+        ):
+            self.onChange.notify(StatChangeEvent(
+                self.parent, attrName, oldVal, value
+            ))
+
+    def get(self, statName):
+        """Looks up a stat value by its name, as it appears in the keys
+        of the `statAbbrs` property.
+
+        Args:
+            statName (str): A stat name, as it appears in the keys of
+                the `statAbbrs` property.
+
+        Returns:
+            int or float: The value of the given stat.
+
+        """
+        return getattr(self, self.statAbbrs[statName])
+
+    def set(self, statName, value):
+        """Sets a stat value by its name, as it appears in the keys of
+        the `statAbbrs` property.
+
+        Args:
+            statName (str): A stat name, as it appears in the keys of
+                the `statAbbrs` property.
+            value (int or float): The value to assign.
+
+        """
+        setattr(self, self.statAbbrs[statName], value)
+
+    def update(self, statDict):
+        """Sets the stat attributes to those contained in the given stat
+        dictionary.
+
+        Args:
+            statDict (dict): {`str`:`int or float`} A dictionary mapping
+                stat names to numerical values. The keys of `statDict`
+                should match the keys of the `statAbbrs` attribute.
+
+        """
+        for statName, statVal in statDict.items():
+            setattr(self, self.statAbbrs[statName], statVal)
 
     def __add__(self, other):
-        result = self.__class__()
-        for statAbbr in self.statAbbrs:
+        result = self.__class__(self.statAbbrs)
+        for statAbbr in self.statAbbrs.values():
             statSum = getattr(self, statAbbr) + getattr(other, statAbbr)
             setattr(result, statAbbr, statSum)
         return result
 
     def __mul__(self, other):
-        result = self.__class__()
-        for statAbbr in self.statAbbrs:
+        result = self.__class__(self.statAbbrs)
+        for statAbbr in self.statAbbrs.values():
             statProd = getattr(self, statAbbr) * getattr(other, statAbbr)
             setattr(result, statAbbr, statProd)
         return result
 
     def __repr__(self):
         return 'StatObject({})'.format(formatDict({
-            statAbbr: getattr(self, statAbbr) for statAbbr in self.statAbbrs
+            statAbbr: getattr(self, statAbbr)
+            for statAbbr in self.statAbbrs.values()
         }))
 
 class Stats(StatObject):
     """Stores the basic stats in STL.
 
-    Attributes:
-        statAbbrs (iterable of str): The values of `STAT_ABBREVIATIONS`,
-            which are abbreviations for the keys of `GSBaseStat`.
+    A subclass of `StatObject` whose `statAbbrs` property is set to
+    `STAT_ABBREVIATIONS`.
 
     """
 
-    def __init__(self, initDict=None):
-        """If a dictionary of stat values is provided, the constructor
-        initializes the instance with these values.
-
-        Args:
-            initDict (dict): {`str`:`int or float`} A dictionary mapping
-                the stat names in the keys of `STAT_ABBREVIATIONS` to
-                numerical values.
-
-        """
-        StatObject.__init__(self)
-        self.statAbbrs = STAT_ABBREVIATIONS.values()
-        if initDict is None:
-            initDict = {statName: 0 for statName in STAT_ABBREVIATIONS}
-        self.update(initDict)
+    def __init__(self, parent=None, initDict=None):
+        StatObject.__init__(self, STAT_ABBREVIATIONS.copy(), parent, initDict)
 
     @property
     def power(self):
@@ -87,54 +197,15 @@ class Stats(StatObject):
             powerDelta += POWER_GRADIENT[statName] * statVal
         return powerDelta
 
-    def get(self, statName):
-        """Looks up a stat value by its full name, as it appears in the
-        keys of `STAT_ABBREVIATIONS`.
-
-        Args:
-            statName (str): A stat name, as it appears in the keys of
-                `STAT_ABBREVIATIONS`.
-
-        Returns:
-            int or float: The value of the given stat.
-
-        """
-        return getattr(self, STAT_ABBREVIATIONS[statName])
-
-    def set(self, statName, value):
-        """Sets a stat value by its full name, as it appears in the keys
-        of `STAT_ABBREVIATIONS`.
-
-        Args:
-            statName (str): A stat name, as it appears in the keys of
-                `STAT_ABBREVIATIONS`.
-            value (int or float): The value to assign.
-
-        """
-        setattr(self, STAT_ABBREVIATIONS[statName], value)
-
-    def update(self, statDict):
-        """Sets the stat attributes to those contained in the given stat
-        dictionary.
-
-        Args:
-            statDict (dict): {`str`:`int or float`} A dictionary mapping
-                the stat names in the keys of `STAT_ABBREVIATIONS` to
-                numerical values.
-
-        """
-        for statName, statVal in statDict.items():
-            setattr(self, STAT_ABBREVIATIONS[statName], statVal)
-
 class PartEffects(StatObject):
     """Stats describing effects of particles in STL.
 
-    Attributes:
-        statAbbrs (iterable of str): The values of `PART_EFFECTS`.
+    A subclass of `StatObject` whose `statAbbrs` property is set to
+    `PART_EFFECTS`.
 
     """
 
-    def __init__(self, initDict=None):
+    def __init__(self, parent=None, initDict=None):
         """If a dictionary of stat values is provided, the constructor
         initializes the instance with these values.
 
@@ -144,47 +215,4 @@ class PartEffects(StatObject):
                 numerical values.
 
         """
-        StatObject.__init__(self)
-        self.statAbbrs = PART_EFFECTS.values()
-        if initDict is None:
-            initDict = {statName: 0 for statName in PART_EFFECTS}
-        self.update(initDict)
-
-    def get(self, statName):
-        """Looks up a stat value by its full name, as it appears in the
-        keys of `PART_EFFECTS`.
-
-        Args:
-            statName (str): A stat name, as it appears in the keys of
-                `PART_EFFECTS`.
-
-        Returns:
-            int or float: The value of the given stat.
-
-        """
-        return getattr(self, PART_EFFECTS[statName])
-
-    def set(self, statName, value):
-        """Sets a stat value by its full name, as it appears in the keys
-        of `PART_EFFECTS`.
-
-        Args:
-            statName (str): A stat name, as it appears in the keys of
-                `PART_EFFECTS`.
-            value (int or float): The value to assign.
-
-        """
-        setattr(self, PART_EFFECTS[statName], value)
-
-    def update(self, statDict):
-        """Sets the stat attributes to those contained in the given stat
-        dictionary.
-
-        Args:
-            statDict (dict): {`str`:`int or float`} A dictionary mapping
-                the stat names in the keys of `PART_EFFECTS` to
-                numerical values.
-
-        """
-        for statName, statVal in statDict.items():
-            setattr(self, PART_EFFECTS[statName], statVal)
+        StatObject.__init__(self, PART_EFFECTS, parent, initDict)